webApp/src/app/companies/services/proyectos.service.ts
import { HttpClient } from '@angular/common/http';
import { Injectable } from '@angular/core';
import { Observable } from 'rxjs';
import { environment } from 'src/environments/environment.development';
import { Proyecto } from '../models/proyecto';

@Injectable({
  providedIn: 'root'
})
export class ProyectosService {

  constructor(
    private httpClient: HttpClient
  ) { }

  listProyectos(empresaId: number): Observable<any>{
    let baseUrl: string = environment.HOST_EMP + 'empresa/' + empresaId + '/proyecto';
    return this.httpClient.get<Proyecto>(baseUrl);
  }
}
